+++ Assets/Scripts/Models/UserModel.cs
@@ -5,6 +5,6 @@ public class UserModel
 {
     public string id;
     public string username;
-    public float money;
+    public decimal money;
     public string partTime;
 }
